+++ agent/src/main/java/com/teamscale/jacoco/agent/AgentBase.java
@@ -4,19 +4,27 @@ import com.teamscale.client.HttpUtils;
 import com.teamscale.jacoco.agent.options.AgentOptionParseException;
 import com.teamscale.jacoco.agent.options.AgentOptions;
 import com.teamscale.jacoco.agent.options.AgentOptionsParser;
+import com.teamscale.jacoco.agent.options.FilePatternResolver;
 import com.teamscale.jacoco.agent.options.JacocoAgentBuilder;
 import com.teamscale.jacoco.agent.util.LoggingUtils;
+import com.teamscale.jacoco.agent.util.LoggingUtils.LoggingResources;
+
+import org.conqat.lib.commons.filesystem.FileSystemUtils;
 import org.jacoco.agent.rt.RT;
 import org.slf4j.Logger;
 import spark.Service;
 
+import java.io.File;
+import java.io.IOException;
 import java.lang.instrument.Instrumentation;
+import java.util.Optional;
 
 /**
- * Base class for agent implementations. Handles logger shutdown, store creation and instantiation of the {@link
- * JacocoRuntimeController}.
+ * Base class for agent implementations. Handles logger shutdown, store creation
+ * and instantiation of the {@link JacocoRuntimeController}.
  * <p>
- * Subclasses must handle dumping onto disk and uploading via the configured uploader.
+ * Subclasses must handle dumping onto disk and uploading via the configured
+ * uploader.
  */
 public abstract class AgentBase {
 
@@ -50,32 +58,41 @@ public abstract class AgentBase {
 	}
 
 	/**
-	 * Starts the http server, which waits for information about started and finished tests.
+	 * Starts the http server, which waits for information about started and
+	 * finished tests.
 	 */
 	private void initServer() {
 		logger.info("Listening for test events on port {}.", options.getHttpServerPort());
 		spark.port(options.getHttpServerPort());
 
 		initServerEndpoints(spark);
-		// this is needed during our tests which will try to access the API directly after creating an agent
+		// this is needed during our tests which will try to access the API
+		// directly after creating an agent
 		spark.awaitInitialization();
 	}
 
 	/** Adds the endpoints that are available in the implemented mode. */
 	protected abstract void initServerEndpoints(Service spark);
 
-	/** Called by the actual premain method once the agent is isolated from the rest of the application. */
+	/**
+	 * Called by the actual premain method once the agent is isolated from the
+	 * rest of the application.
+	 */
 	public static void premain(String options, Instrumentation instrumentation) throws Exception {
 		AgentOptions agentOptions;
 		DelayedLogger delayedLogger = new DelayedLogger();
 		try {
 			agentOptions = AgentOptionsParser.parse(options, delayedLogger);
 		} catch (AgentOptionParseException e) {
-			try (LoggingUtils.LoggingResources ignored = LoggingUtils.initializeDefaultLogging()) {
-				Logger logger = LoggingUtils.getLogger(PreMain.class);
-				delayedLogger.logTo(logger);
-				logger.error("Failed to parse agent options: " + e.getMessage(), e);
+			try (LoggingUtils.LoggingResources ignored = initializeFallbackLogging(options, delayedLogger)) {
+				delayedLogger.error("Failed to parse agent options: " + e.getMessage(), e);
 				System.err.println("Failed to parse agent options: " + e.getMessage());
+
+				// we perform actual logging output after writing to console to
+				// ensure the console is reached even in case of logging issues
+				// (see TS-23151). We use the Agent class here (same as below)
+				Logger logger = LoggingUtils.getLogger(Agent.class);
+				delayedLogger.logTo(logger);
 				throw e;
 			}
 		}
@@ -89,25 +106,79 @@ public abstract class AgentBase {
 
 		logger.info("Starting JaCoCo's agent");
 		JacocoAgentBuilder agentBuilder = new JacocoAgentBuilder(agentOptions);
-		org.jacoco.agent.rt.internal_43f5073.PreMain
-				.premain(agentBuilder.createJacocoAgentOptions(), instrumentation);
+		org.jacoco.agent.rt.internal_43f5073.PreMain.premain(agentBuilder.createJacocoAgentOptions(), instrumentation);
 
 		AgentBase agent = agentBuilder.createAgent(instrumentation);
 		agent.registerShutdownHook();
 	}
 
 	/**
-	 * Registers a shutdown hook that stops the timer and dumps coverage a final time.
+	 * Initializes fallback logging in case of an error during the parsing of
+	 * the options to {@link #premain(String, Instrumentation)} (see TS-23151).
+	 * This tries to extract the logging configuration and use this and falls
+	 * back to the default logger.
+	 */
+	private static LoggingResources initializeFallbackLogging(String premainOptions, DelayedLogger delayedLogger) {
+		for (String optionPart : premainOptions.split(",")) {
+			if (optionPart.startsWith(AgentOptionsParser.LOGGING_CONFIG_OPTION + "=")) {
+				return createFallbackLoggerFromConfig(optionPart.split("=", 2)[1], delayedLogger);
+			}
+			
+			if (optionPart.startsWith(AgentOptionsParser.CONFIG_FILE_OPTION + "=")) {
+				String configFileValue = optionPart.split("=", 2)[1];
+				Optional<String> loggingConfigLine = Optional.empty();
+				try {
+					File configFile = new FilePatternResolver(delayedLogger)
+							.parsePath(AgentOptionsParser.CONFIG_FILE_OPTION, configFileValue).toFile();
+					loggingConfigLine = FileSystemUtils.readLinesUTF8(configFile).stream()
+							.filter(line -> line.startsWith(AgentOptionsParser.LOGGING_CONFIG_OPTION + "="))
+							.findFirst();
+				} catch (IOException | AgentOptionParseException e) {
+					delayedLogger.error("Failed to load configuration from " + configFileValue + ": " + e.getMessage(),
+							e);
+				}
+				if (loggingConfigLine.isPresent()) {
+					return createFallbackLoggerFromConfig(loggingConfigLine.get().split("=", 2)[1], delayedLogger);
+				}
+			}
+		}
+
+		return LoggingUtils.initializeDefaultLogging();
+	}
+
+	/** Creates a fallback logger using the given config file. */
+	private static LoggingResources createFallbackLoggerFromConfig(String configLocation, DelayedLogger delayedLogger) {
+		try {
+			return LoggingUtils.initializeLogging(new FilePatternResolver(delayedLogger)
+					.parsePath(AgentOptionsParser.LOGGING_CONFIG_OPTION, configLocation));
+		} catch (IOException | AgentOptionParseException e) {
+			String message = "Failed to load log configuration from location " + configLocation + ": " + e.getMessage();
+			delayedLogger.error(message, e);
+			// output the message to console as well, as this might
+			// otherwise not make it to the user
+			System.err.println(message);
+			return LoggingUtils.initializeDefaultLogging();
+		}
+	}
+
+	/**
+	 * Registers a shutdown hook that stops the timer and dumps coverage a final
+	 * time.
 	 */
 	private void registerShutdownHook() {
 		Runtime.getRuntime().addShutdownHook(new Thread(() -> {
-			if (options.getHttpServerPort() != null) {
-				spark.stop();
-			}
+			stopServer();
 			prepareShutdown();
 			logger.info("CQSE JaCoCo agent successfully shut down.");
 			loggingResources.close();
 		}));
+	}
+
+	/** Stop the http server if it's running */
+	void stopServer() {
+		if (options.getHttpServerPort() != null) {
+			spark.stop();
+		}
 	}
 
 	/** Called when the shutdown hook is triggered. */