+++ agent/src/main/java/com/teamscale/jacoco/agent/upload/teamscale/TeamscaleUploader.java
@@ -68,7 +68,7 @@ public class TeamscaleUploader implements IUploader {
 					teamscaleServer.revision,
 					teamscaleServer.partition,
 					EReportFormat.JACOCO,
-					teamscaleServer.message,
+					teamscaleServer.getMessage(),
 					coverageFile.createFormRequestBody()
 			);
 			return true;